requirements.txt
keras==2.7.0
numpy==1.20.1
Pillow==8.4.0
streamlit==1.2.0
